Move generator package to be an internal one
