+++ intermediate/link_chain.go
@@ -2,8 +2,8 @@ package intermediate
 
 import (
 	"errors"
-	"github.com/halprin/rangechain/generator"
 	"github.com/halprin/rangechain/helper"
+	"github.com/halprin/rangechain/internal/generator"
 	"sort"
 )
