+++ intermediate/link_termination_test.go
@@ -3,8 +3,8 @@ package intermediate
 
 import (
 	"errors"
-	"github.com/halprin/rangechain/generator"
 	"github.com/halprin/rangechain/helper"
+	"github.com/halprin/rangechain/internal/generator"
 	"github.com/stretchr/testify/assert"
 	"sync"
 	"testing"
